api/pyproject.toml
[tool.poetry]
name = "discordscrapebotapi"
version = "0.1.0"
description = ""
authors = ["Ethan Christensen <ethanlchristensen@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
django-tables2 = "^2.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
